pyproject.toml
[tool.poetry]
name = "aiogramtemplate"
version = "0.1.0"
description = ""
authors = ["Stanislav Lukyanov <someemail>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.7.0"
python-dotenv = "^1.0.1"
black = "^24.4.2"
pydantic-settings = "^2.3.0"
fluent-runtime = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
